refactor: refactor card and products
since same card is used in product listing and wishlist page, refactored card to handle functions inside instead of in parent component
