+++ src/Components/Card/Card.jsx
@@ -1,21 +1,39 @@
-import React from 'react';
+import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { useAuth, useCart } from '../../Context';
+import { addToCartHandler } from '../../Utils';
 import styles from './Card.module.css';
 
-const Card = ({
-	discount,
-	badge,
-	categoryName,
-	image,
-	price,
-	rating,
-	title,
-	cartBtnHandler,
-	_id,
-	checkCartStatus,
-	cartBtnLoader,
-}) => {
+const Card = ({ discount, badge, categoryName, image, price, rating, title, _id, products }) => {
+	const [cartBtnLoader, setCartBtnLoader] = useState(false);
+
 	const navigate = useNavigate();
+
+	const { authState } = useAuth();
+
+	const { cartState, cartDispatch } = useCart();
+
+	const checkCartStatus = (_id) => {
+		const itemInCart = cartState.cart.find((item) => item._id === _id);
+		return itemInCart ? 'Go to Cart' : 'Add to Cart';
+	};
+
+	const cartBtnHandler = async (_id) => {
+		setCartBtnLoader(true);
+		const product = products.find((product) => product._id === _id);
+		if (authState.token) {
+			const response = await addToCartHandler(product, authState.token);
+
+			if (response.status === 201) {
+				cartDispatch({ type: 'ADD_TO_CART', payload: response.data.cart });
+				setCartBtnLoader(false);
+			}
+		} else {
+			alert('You are not logged in');
+			navigate('/login');
+		}
+	};
+
 	return (
 		<article className={`card product-card card-shadow ${styles.card}`}>
 			<img src={image} alt="card image 1" className={`card-img ${styles.cardImg}`} />
